+++ code/logic/fossil/test/mocking.h
